+++ sgl-router/src/protocols/mod.rs
@@ -1,6 +1,27 @@
 // Protocol definitions and validation for various LLM APIs
 // This module provides a structured approach to handling different API protocols
 
-pub mod spec;
+pub mod chat;
+pub mod common;
+pub mod completion;
+pub mod embedding;
+pub mod generate;
+pub mod rerank;
+pub mod sampling_params;
 pub mod validated;
 pub mod worker_spec;
+
+// Keep spec.rs for now to avoid breaking changes during migration
+// TODO: This can be removed once all imports are updated
+pub mod spec;
+
+// Re-export all public types for backward compatibility
+pub use chat::*;
+pub use common::*;
+pub use completion::*;
+pub use embedding::*;
+pub use generate::*;
+pub use rerank::*;
+pub use sampling_params::*;
+pub use validated::*;
+pub use worker_spec::*;
